Doing it right
Sigh

+++ Content.Server/Station/Systems/StationSpawningSystem.cs
@@ -160,7 +160,7 @@ public sealed class StationSpawningSystem : SharedStationSpawningSystem
             if (profile.FlavorText != "" && _configurationManager.GetCVar(CCVars.FlavorText))
                 EnsureComp<DetailExaminableComponent>(entity.Value).Content = profile.FlavorText;
             var bank = EnsureComp<BankAccountComponent>(entity.Value);
-            bank.Balance = profile.BankBalance;
+            bank.Balance = (int)profile.BankBalance;
         }
 
         DoJobSpecials(job, entity.Value);
